creds_manager/requirements.txt
pbr
enum34
eventlet
keystoneauth1
keystonemiddleware
greenlet
MySQL-python
pymysql
oslo.config
oslo.concurrency
oslo.db
oslo.log
oslo.messaging
oslo.middleware
oslo.policy
oslo.service
SQLAlchemy
sqlalchemy-migrate
webob
cryptography
